+++ CHANGELOG.md
@@ -15,6 +15,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Moved definitions file loading to post-initialization
 
+### Fixed
+
+- Fixed regression where diagnostics are not cleared when you close an ignored file
+
 ## [1.5.2] - 2022-06-22
 
 ### Changed